+++ packages/core/__tests__/standardize.spec.ts
@@ -12,4 +12,8 @@ describe('encode', () => {
   it('add now', () => {
     expect(standardize('+d')).toEqual('now+d');
   });
+
+  it('show 1 when set displayOne to true', () => {
+    expect(standardize('+d/h-m', { displayOne: true })).toBe('now+1d/h-1m');
+  });
 })